+++ src/common/hooks/search/use-recent-items.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { useRecoilState } from 'recoil';
-import { searchRecentlyViewedItemsState } from '@store/search';
+import { searchRecentlyViewedItemsState } from '@store/recoil/search';
 import { useApiServer } from '@common/hooks/use-api';
 
 export const useRecentlyViewedItems = () => {
