+++ src/common/hooks/search/use-search-focus.tsx
@@ -1,4 +1,4 @@
-import { searchFocusedState } from '@store/search';
+import { searchFocusedState } from '@store/recoil/search';
 import { useRecoilFocus } from '@common/hooks/use-recoil-focus';
 
 export const useSearchFocus = () => {
