+++ src/common/hooks/use-api.tsx
@@ -1,5 +1,5 @@
 import { useRecoilValue } from 'recoil';
-import { networkCurrentUrlSelector } from '@store/network';
+import { networkCurrentUrlSelector } from '@store/recoil/network';
 
 export const useApiServer = (): string => {
   const network = useRecoilValue(networkCurrentUrlSelector);
