+++ src/common/hooks/use-unlocking-state.ts
@@ -1,5 +1,5 @@
 import { useRecoilState } from 'recoil';
-import { unlockingState } from '@store/unlocking';
+import { unlockingState } from '@store/recoil/unlocking';
 
 export const useUnlockingState = () => {
   return useRecoilState(unlockingState);
